Use useUserVar from Provider to give context to different areas (ex. testing)
