+++ frontend/src/CreateUserPage.js
@@ -5,7 +5,7 @@ import { object, string } from 'yup'
 import { Link as RouteLink, useHistory, useParams } from 'react-router-dom'
 import { Formik } from 'formik'
 import { SIGN_UP } from './graphql/mutations'
-import { useUserVar } from './useUserVar'
+import { useUserVar } from './UserState'
 import EmailField from './EmailField'
 import PasswordConfirmation from './PasswordConfirmation'
 import LanguageSelect from './LanguageSelect'
