+++ frontend/src/QRcodePage.js
@@ -3,7 +3,7 @@ import { Trans } from '@lingui/macro'
 import { Box, Button, Stack, Text } from '@chakra-ui/core'
 import { useHistory } from 'react-router-dom'
 import { useQuery } from '@apollo/client'
-import { useUserVar } from './useUserVar'
+import { useUserVar } from './UserState'
 import { GENERATE_OTP_URL } from './graphql/queries'
 import QRCode from 'qrcode.react'
 import { ErrorBoundary } from 'react-error-boundary'
