+++ frontend/src/TwoFactorAuthenticatePage.js
@@ -6,7 +6,7 @@ import { Box, Heading, Stack, useToast } from '@chakra-ui/core'
 import { useHistory, useLocation, useParams } from 'react-router-dom'
 import { useMutation } from '@apollo/client'
 import { Formik } from 'formik'
-import { useUserVar } from './useUserVar'
+import { useUserVar } from './UserState'
 import { AUTHENTICATE } from './graphql/mutations'
 import AuthenticateField from './AuthenticateField'
 import { fieldRequirements } from './fieldRequirements'
