+++ frontend/src/__tests__/DomainsPage.test.js
@@ -6,10 +6,11 @@ import { MockedProvider } from '@apollo/client/testing'
 import { PAGINATED_DOMAINS } from '../graphql/queries'
 import { I18nProvider } from '@lingui/react'
 import { setupI18n } from '@lingui/core'
-import { UserStateProvider } from '../UserState'
+import { UserVarProvider } from '../UserState'
 import { createCache } from '../client'
 import DomainsPage from '../DomainsPage'
 import { createMemoryHistory } from 'history'
+import { makeVar } from '@apollo/client'
 
 const i18n = setupI18n({
   locale: 'en',
@@ -177,8 +178,12 @@ describe('<DomainsPage />', () => {
     it('displays a list of domains', async () => {
       const { queryByText } = render(
         <MockedProvider mocks={mocks} cache={createCache()}>
-          <UserStateProvider
-            initialState={{ userName: null, jwt: null, tfaSendMethod: null }}
+          <UserVarProvider
+            userVar={makeVar({
+              jwt: null,
+              tfaSendMethod: null,
+              userName: null,
+            })}
           >
             <ThemeProvider theme={theme}>
               <I18nProvider i18n={i18n}>
@@ -187,7 +192,7 @@ describe('<DomainsPage />', () => {
                 </MemoryRouter>
               </I18nProvider>
             </ThemeProvider>
-          </UserStateProvider>
+          </UserVarProvider>
         </MockedProvider>,
       )
 
@@ -199,8 +204,12 @@ describe('<DomainsPage />', () => {
     it('handles an empty list of domains', async () => {
       const { queryByText } = render(
         <MockedProvider mocks={empty} cache={createCache()}>
-          <UserStateProvider
-            initialState={{ userName: null, jwt: null, tfaSendMethod: null }}
+          <UserVarProvider
+            userVar={makeVar({
+              jwt: null,
+              tfaSendMethod: null,
+              userName: null,
+            })}
           >
             <ThemeProvider theme={theme}>
               <I18nProvider i18n={i18n}>
@@ -209,7 +218,7 @@ describe('<DomainsPage />', () => {
                 </MemoryRouter>
               </I18nProvider>
             </ThemeProvider>
-          </UserStateProvider>
+          </UserVarProvider>
         </MockedProvider>,
       )
 
@@ -226,8 +235,12 @@ describe('<DomainsPage />', () => {
       it('takes user to DMARC Report page', async () => {
         const { getAllByText } = render(
           <MockedProvider mocks={mocks} cache={createCache()}>
-            <UserStateProvider
-              initialState={{ userName: null, jwt: null, tfaSendMethod: null }}
+            <UserVarProvider
+              userVar={makeVar({
+                jwt: null,
+                tfaSendMethod: null,
+                userName: null,
+              })}
             >
               <ThemeProvider theme={theme}>
                 <I18nProvider i18n={i18n}>
@@ -240,7 +253,7 @@ describe('<DomainsPage />', () => {
                   </MemoryRouter>
                 </I18nProvider>
               </ThemeProvider>
-            </UserStateProvider>
+            </UserVarProvider>
           </MockedProvider>,
         )
 
@@ -256,8 +269,12 @@ describe('<DomainsPage />', () => {
       it('takes user to Guidance page', async () => {
         const { getAllByText } = render(
           <MockedProvider mocks={mocks} cache={createCache()}>
-            <UserStateProvider
-              initialState={{ userName: null, jwt: null, tfaSendMethod: null }}
+            <UserVarProvider
+              userVar={makeVar({
+                jwt: null,
+                tfaSendMethod: null,
+                userName: null,
+              })}
             >
               <ThemeProvider theme={theme}>
                 <I18nProvider i18n={i18n}>
@@ -270,7 +287,7 @@ describe('<DomainsPage />', () => {
                   </MemoryRouter>
                 </I18nProvider>
               </ThemeProvider>
-            </UserStateProvider>
+            </UserVarProvider>
           </MockedProvider>,
         )
 
@@ -287,12 +304,12 @@ describe('<DomainsPage />', () => {
         it('correctly filters results', async () => {
           const { getByPlaceholderText, queryByText } = render(
             <MockedProvider mocks={mocks} cache={createCache()}>
-              <UserStateProvider
-                initialState={{
-                  userName: null,
+              <UserVarProvider
+                userVar={makeVar({
                   jwt: null,
                   tfaSendMethod: null,
-                }}
+                  userName: null,
+                })}
               >
                 <ThemeProvider theme={theme}>
                   <I18nProvider i18n={i18n}>
@@ -304,7 +321,7 @@ describe('<DomainsPage />', () => {
                     </MemoryRouter>
                   </I18nProvider>
                 </ThemeProvider>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>,
           )
 
@@ -329,12 +346,12 @@ describe('<DomainsPage />', () => {
         it('changes sorting order', async () => {
           const { getByTestId } = render(
             <MockedProvider mocks={mocks} cache={createCache()}>
-              <UserStateProvider
-                initialState={{
-                  userName: null,
+              <UserVarProvider
+                userVar={makeVar({
                   jwt: null,
                   tfaSendMethod: null,
-                }}
+                  userName: null,
+                })}
               >
                 <ThemeProvider theme={theme}>
                   <I18nProvider i18n={i18n}>
@@ -346,7 +363,7 @@ describe('<DomainsPage />', () => {
                     </MemoryRouter>
                   </I18nProvider>
                 </ThemeProvider>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>,
           )
