+++ frontend/src/__tests__/EditableUserDisplayName.test.js
@@ -4,11 +4,12 @@ import { theme, ThemeProvider } from '@chakra-ui/core'
 import EditableUserDisplayName from '../EditableUserDisplayName'
 import { I18nProvider } from '@lingui/react'
 import { setupI18n } from '@lingui/core'
-import { UserStateProvider } from '../UserState'
+import { UserVarProvider } from '../UserState'
 import { MemoryRouter } from 'react-router-dom'
 import { MockedProvider } from '@apollo/client/testing'
 import { fireEvent } from '@testing-library/dom'
 import { UPDATE_USER_PROFILE } from '../graphql/mutations'
+import { makeVar } from '@apollo/client'
 
 const i18n = setupI18n({
   locale: 'en',
@@ -24,12 +25,8 @@ describe('<EditableUserDisplayName>', () => {
   it('renders', async () => {
     const { getByText } = render(
       <MockedProvider addTypename={false}>
-        <UserStateProvider
-          initialState={{
-            userName: 'testUserName@email.com',
-            jwt: 'string',
-            tfaSendMethod: false,
-          }}
+        <UserVarProvider
+          userVar={makeVar({ jwt: null, tfaSendMethod: null, userName: null })}
         >
           <MemoryRouter initialEntries={['/']}>
             <I18nProvider i18n={i18n}>
@@ -38,7 +35,7 @@ describe('<EditableUserDisplayName>', () => {
               </ThemeProvider>
             </I18nProvider>
           </MemoryRouter>
-        </UserStateProvider>
+        </UserVarProvider>
       </MockedProvider>,
     )
     await waitFor(() => expect(getByText(/Edit/i)).toBeInTheDocument())
@@ -47,12 +44,12 @@ describe('<EditableUserDisplayName>', () => {
     it('opens the modal', async () => {
       const { getByText } = render(
         <MockedProvider addTypename={false}>
-          <UserStateProvider
-            initialState={{
-              userName: 'testUserName@email.com',
-              jwt: 'string',
-              tfaSendMethod: false,
-            }}
+          <UserVarProvider
+            userVar={makeVar({
+              jwt: null,
+              tfaSendMethod: null,
+              userName: null,
+            })}
           >
             <MemoryRouter initialEntries={['/']}>
               <I18nProvider i18n={i18n}>
@@ -61,7 +58,7 @@ describe('<EditableUserDisplayName>', () => {
                 </ThemeProvider>
               </I18nProvider>
             </MemoryRouter>
-          </UserStateProvider>
+          </UserVarProvider>
         </MockedProvider>,
       )
       const editButton = getByText(/Edit/i)
@@ -78,12 +75,12 @@ describe('<EditableUserDisplayName>', () => {
         it('displays field error', async () => {
           const { getByText } = render(
             <MockedProvider addTypename={false}>
-              <UserStateProvider
-                initialState={{
-                  userName: 'testUserName@email.com',
-                  jwt: 'string',
-                  tfaSendMethod: false,
-                }}
+              <UserVarProvider
+                userVar={makeVar({
+                  jwt: null,
+                  tfaSendMethod: null,
+                  userName: null,
+                })}
               >
                 <MemoryRouter initialEntries={['/']}>
                   <I18nProvider i18n={i18n}>
@@ -92,7 +89,7 @@ describe('<EditableUserDisplayName>', () => {
                     </ThemeProvider>
                   </I18nProvider>
                 </MemoryRouter>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>,
           )
           const editButton = getByText(/Edit/i)
@@ -143,12 +140,12 @@ describe('<EditableUserDisplayName>', () => {
 
           const { queryByText, getByText, getByLabelText } = render(
             <MockedProvider addTypename={false} mocks={mocks}>
-              <UserStateProvider
-                initialState={{
-                  userName: 'testUserName@email.com',
-                  jwt: 'string',
-                  tfaSendMethod: false,
-                }}
+              <UserVarProvider
+                userVar={makeVar({
+                  jwt: null,
+                  tfaSendMethod: null,
+                  userName: null,
+                })}
               >
                 <MemoryRouter initialEntries={['/']}>
                   <I18nProvider i18n={i18n}>
@@ -157,7 +154,7 @@ describe('<EditableUserDisplayName>', () => {
                     </ThemeProvider>
                   </I18nProvider>
                 </MemoryRouter>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>,
           )
           const editButton = getByText(/Edit/i)
@@ -204,12 +201,12 @@ describe('<EditableUserDisplayName>', () => {
 
           const { queryByText, getByText, getByLabelText } = render(
             <MockedProvider addTypename={false} mocks={mocks}>
-              <UserStateProvider
-                initialState={{
-                  userName: 'testUserName@email.com',
-                  jwt: 'string',
-                  tfaSendMethod: false,
-                }}
+              <UserVarProvider
+                userVar={makeVar({
+                  jwt: null,
+                  tfaSendMethod: null,
+                  userName: null,
+                })}
               >
                 <MemoryRouter initialEntries={['/']}>
                   <I18nProvider i18n={i18n}>
@@ -218,7 +215,7 @@ describe('<EditableUserDisplayName>', () => {
                     </ThemeProvider>
                   </I18nProvider>
                 </MemoryRouter>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>,
           )
           const editButton = getByText(/Edit/i)
