+++ frontend/src/__tests__/EditableUserEmail.test.js
@@ -4,11 +4,12 @@ import { theme, ThemeProvider } from '@chakra-ui/core'
 import EditableUserEmail from '../EditableUserEmail'
 import { I18nProvider } from '@lingui/react'
 import { setupI18n } from '@lingui/core'
-import { UserStateProvider } from '../UserState'
+import { UserVarProvider } from '../UserState'
 import { MemoryRouter } from 'react-router-dom'
 import { MockedProvider } from '@apollo/client/testing'
 import { fireEvent } from '@testing-library/dom'
 import { UPDATE_USER_PROFILE } from '../graphql/mutations'
+import { makeVar } from '@apollo/client'
 
 const i18n = setupI18n({
   locale: 'en',
@@ -24,12 +25,8 @@ describe('<EditableUserEmail />', () => {
   it('renders', async () => {
     const { getByText } = render(
       <MockedProvider addTypename={false}>
-        <UserStateProvider
-          initialState={{
-            userName: 'testUserName@email.com',
-            jwt: 'string',
-            tfaSendMethod: false,
-          }}
+        <UserVarProvider
+          userVar={makeVar({ jwt: null, tfaSendMethod: null, userName: null })}
         >
           <MemoryRouter initialEntries={['/']}>
             <I18nProvider i18n={i18n}>
@@ -38,7 +35,7 @@ describe('<EditableUserEmail />', () => {
               </ThemeProvider>
             </I18nProvider>
           </MemoryRouter>
-        </UserStateProvider>
+        </UserVarProvider>
       </MockedProvider>,
     )
     await waitFor(() => expect(getByText(/Edit/i)).toBeInTheDocument())
@@ -47,12 +44,12 @@ describe('<EditableUserEmail />', () => {
     it('opens the modal', async () => {
       const { getByText } = render(
         <MockedProvider addTypename={false}>
-          <UserStateProvider
-            initialState={{
-              userName: 'testUserName@email.com',
-              jwt: 'string',
-              tfaSendMethod: false,
-            }}
+          <UserVarProvider
+            userVar={makeVar({
+              jwt: null,
+              tfaSendMethod: null,
+              userName: null,
+            })}
           >
             <MemoryRouter initialEntries={['/']}>
               <I18nProvider i18n={i18n}>
@@ -61,7 +58,7 @@ describe('<EditableUserEmail />', () => {
                 </ThemeProvider>
               </I18nProvider>
             </MemoryRouter>
-          </UserStateProvider>
+          </UserVarProvider>
         </MockedProvider>,
       )
       const editButton = getByText(/Edit/i)
@@ -78,12 +75,12 @@ describe('<EditableUserEmail />', () => {
         it('displays field error', async () => {
           const { getByText } = render(
             <MockedProvider addTypename={false}>
-              <UserStateProvider
-                initialState={{
-                  userName: 'testUserName@email.com',
-                  jwt: 'string',
-                  tfaSendMethod: false,
-                }}
+              <UserVarProvider
+                userVar={makeVar({
+                  jwt: null,
+                  tfaSendMethod: null,
+                  userName: null,
+                })}
               >
                 <MemoryRouter initialEntries={['/']}>
                   <I18nProvider i18n={i18n}>
@@ -92,7 +89,7 @@ describe('<EditableUserEmail />', () => {
                     </ThemeProvider>
                   </I18nProvider>
                 </MemoryRouter>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>,
           )
           const editButton = getByText(/Edit/i)
@@ -141,12 +138,12 @@ describe('<EditableUserEmail />', () => {
 
           const { queryByText, getByText, getByLabelText } = render(
             <MockedProvider addTypename={false} mocks={mocks}>
-              <UserStateProvider
-                initialState={{
-                  userName: 'testUserName@email.com',
-                  jwt: 'string',
-                  tfaSendMethod: false,
-                }}
+              <UserVarProvider
+                userVar={makeVar({
+                  jwt: null,
+                  tfaSendMethod: null,
+                  userName: null,
+                })}
               >
                 <MemoryRouter initialEntries={['/']}>
                   <I18nProvider i18n={i18n}>
@@ -155,7 +152,7 @@ describe('<EditableUserEmail />', () => {
                     </ThemeProvider>
                   </I18nProvider>
                 </MemoryRouter>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>,
           )
           const editButton = getByText(/Edit/i)
@@ -204,12 +201,12 @@ describe('<EditableUserEmail />', () => {
 
           const { queryByText, getByText, getByLabelText } = render(
             <MockedProvider addTypename={false} mocks={mocks}>
-              <UserStateProvider
-                initialState={{
-                  userName: 'testUserName@email.com',
-                  jwt: 'string',
-                  tfaSendMethod: false,
-                }}
+              <UserVarProvider
+                userVar={makeVar({
+                  jwt: null,
+                  tfaSendMethod: null,
+                  userName: null,
+                })}
               >
                 <MemoryRouter initialEntries={['/']}>
                   <I18nProvider i18n={i18n}>
@@ -218,7 +215,7 @@ describe('<EditableUserEmail />', () => {
                     </ThemeProvider>
                   </I18nProvider>
                 </MemoryRouter>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>,
           )
           const editButton = getByText(/Edit/i)
