+++ frontend/src/__tests__/OrganizationAffiliations.test.js
@@ -3,12 +3,13 @@ import { render, waitFor } from '@testing-library/react'
 import { MockedProvider } from '@apollo/client/testing'
 import { MemoryRouter, Route } from 'react-router-dom'
 import { theme, ThemeProvider } from '@chakra-ui/core'
-import { UserStateProvider } from '../UserState'
+import { UserVarProvider } from '../UserState'
 import { PAGINATED_ORG_AFFILIATIONS } from '../graphql/queries'
 import { I18nProvider } from '@lingui/react'
 import { setupI18n } from '@lingui/core'
 import { OrganizationAffiliations } from '../OrganizationAffiliations'
 import matchMediaPolyfill from 'mq-polyfill'
+import { makeVar } from '@apollo/client'
 
 const i18n = setupI18n({
   locale: 'en',
@@ -121,12 +122,12 @@ describe('<OrganizationAffiliations />', () => {
         <ThemeProvider theme={theme}>
           <I18nProvider i18n={i18n}>
             <MockedProvider mocks={mocks} addTypename={false}>
-              <UserStateProvider
-                initialState={{
-                  userName: 'user@example.com',
-                  jwt: 'somestring',
+              <UserVarProvider
+                userVar={makeVar({
+                  jwt: null,
                   tfaSendMethod: null,
-                }}
+                  userName: null,
+                })}
               >
                 <MemoryRouter
                   initialEntries={['/organization/tbs-sct-gc-ca']}
@@ -136,7 +137,7 @@ describe('<OrganizationAffiliations />', () => {
                     <OrganizationAffiliations orgSlug={orgSlug} />
                   </Route>
                 </MemoryRouter>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>
           </I18nProvider>
         </ThemeProvider>,
