+++ frontend/src/__tests__/OrganizationDetails.test.js
@@ -3,12 +3,13 @@ import { render, waitFor } from '@testing-library/react'
 import { MockedProvider } from '@apollo/client/testing'
 import { MemoryRouter, Route } from 'react-router-dom'
 import { theme, ThemeProvider } from '@chakra-ui/core'
-import { UserStateProvider } from '../UserState'
+import { UserVarProvider } from '../UserState'
 import { ORG_DETAILS_PAGE } from '../graphql/queries'
 import { I18nProvider } from '@lingui/react'
 import { setupI18n } from '@lingui/core'
 import OrganizationDetails from '../OrganizationDetails'
 import matchMediaPolyfill from 'mq-polyfill'
+import { makeVar } from '@apollo/client'
 
 const i18n = setupI18n({
   locale: 'en',
@@ -115,12 +116,12 @@ describe('<OrganizationDetails />', () => {
         <ThemeProvider theme={theme}>
           <I18nProvider i18n={i18n}>
             <MockedProvider mocks={mocks} addTypename={false}>
-              <UserStateProvider
-                initialState={{
-                  userName: 'user@example.com',
-                  jwt: 'somestring',
+              <UserVarProvider
+                userVar={makeVar({
+                  jwt: null,
                   tfaSendMethod: null,
-                }}
+                  userName: null,
+                })}
               >
                 <MemoryRouter
                   initialEntries={['/organizations/tbs-sct-gc-ca']}
@@ -130,7 +131,7 @@ describe('<OrganizationDetails />', () => {
                     <OrganizationDetails />
                   </Route>
                 </MemoryRouter>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>
           </I18nProvider>
         </ThemeProvider>,
