+++ frontend/src/__tests__/OrganizationDomains.test.js
@@ -3,12 +3,13 @@ import { render, waitFor } from '@testing-library/react'
 import { MockedProvider } from '@apollo/client/testing'
 import { MemoryRouter, Route } from 'react-router-dom'
 import { theme, ThemeProvider } from '@chakra-ui/core'
-import { UserStateProvider } from '../UserState'
+import { UserVarProvider } from '../UserState'
 import { PAGINATED_ORG_DOMAINS } from '../graphql/queries'
 import { I18nProvider } from '@lingui/react'
 import { setupI18n } from '@lingui/core'
 import { OrganizationDomains } from '../OrganizationDomains'
 import matchMediaPolyfill from 'mq-polyfill'
+import { makeVar } from '@apollo/client'
 
 const i18n = setupI18n({
   locale: 'en',
@@ -108,12 +109,12 @@ describe('<OrganizationDomains />', () => {
         <ThemeProvider theme={theme}>
           <I18nProvider i18n={i18n}>
             <MockedProvider mocks={mocks} addTypename={false}>
-              <UserStateProvider
-                initialState={{
-                  userName: 'user@example.com',
-                  jwt: 'somestring',
+              <UserVarProvider
+                userVar={makeVar({
+                  jwt: null,
                   tfaSendMethod: null,
-                }}
+                  userName: null,
+                })}
               >
                 <MemoryRouter
                   initialEntries={['/organization/tbs-sct-gc-ca']}
@@ -123,7 +124,7 @@ describe('<OrganizationDomains />', () => {
                     <OrganizationDomains orgSlug={orgSlug} />
                   </Route>
                 </MemoryRouter>
-              </UserStateProvider>
+              </UserVarProvider>
             </MockedProvider>
           </I18nProvider>
         </ThemeProvider>,
