+++ frontend/src/__tests__/ScanCard.test.js
@@ -5,9 +5,10 @@ import { render, waitFor } from '@testing-library/react'
 import ScanCard from '../ScanCard'
 import { I18nProvider } from '@lingui/react'
 import { setupI18n } from '@lingui/core'
-import { UserStateProvider } from '../UserState'
+import { UserVarProvider } from '../UserState'
 import { rawDmarcGuidancePageData } from '../fixtures/dmarcGuidancePageData'
 import { MockedProvider } from '@apollo/client/testing'
+import { makeVar } from '@apollo/client'
 
 const i18n = setupI18n({
   locale: 'en',
@@ -35,8 +36,8 @@ describe('<ScanCard />', () => {
   it('renders', async () => {
     const { getAllByText } = render(
       <MockedProvider>
-        <UserStateProvider
-          initialState={{ userName: null, jwt: null, tfaSendMethod: null }}
+        <UserVarProvider
+          userVar={makeVar({ jwt: null, tfaSendMethod: null, userName: null })}
         >
           <ThemeProvider theme={theme}>
             <I18nProvider i18n={i18n}>
@@ -49,7 +50,7 @@ describe('<ScanCard />', () => {
               </MemoryRouter>
             </I18nProvider>
           </ThemeProvider>
-        </UserStateProvider>
+        </UserVarProvider>
       </MockedProvider>,
     )
     await waitFor(() => getAllByText(/Web Scan Results/i))
