+++ frontend/src/__tests__/UserPage.test.js
@@ -6,8 +6,9 @@ import { MemoryRouter } from 'react-router-dom'
 import { theme, ThemeProvider } from '@chakra-ui/core'
 import { I18nProvider } from '@lingui/react'
 import { MockedProvider } from '@apollo/client/testing'
-import { UserStateProvider } from '../UserState'
+import { UserVarProvider } from '../UserState'
 import { QUERY_CURRENT_USER } from '../graphql/queries'
+import { makeVar } from '@apollo/client'
 
 const i18n = setupI18n({
   locale: 'en',
@@ -53,8 +54,8 @@ describe('<UserPage />', () => {
   it('renders without error', async () => {
     const { queryByText } = render(
       <MockedProvider mocks={mocks} addTypename={false}>
-        <UserStateProvider
-          initialState={{ userName, jwt: 'string', tfaSendMethod: null }}
+        <UserVarProvider
+          userVar={makeVar({ jwt: null, tfaSendMethod: null, userName: null })}
         >
           <MemoryRouter initialEntries={['/']}>
             <ThemeProvider theme={theme}>
@@ -63,7 +64,7 @@ describe('<UserPage />', () => {
               </I18nProvider>
             </ThemeProvider>
           </MemoryRouter>
-        </UserStateProvider>
+        </UserVarProvider>
       </MockedProvider>,
     )
     await waitFor(() => expect(queryByText(userName)).toBeInTheDocument())
