+++ frontend/src/index.js
@@ -7,18 +7,21 @@ import App from './App'
 import * as serviceWorker from './serviceWorker'
 import { BrowserRouter as Router } from 'react-router-dom'
 import canada from './theme/canada'
-import { client } from './client'
+import { client, currentUserVar } from './client'
 import { ApolloProvider } from '@apollo/client'
+import { UserVarProvider } from './UserState'
 
 ReactDOM.render(
   <ApolloProvider client={client}>
-    <ThemeProvider theme={canada}>
-      <I18nProvider i18n={i18n}>
-        <Router>
-          <App />
-        </Router>
-      </I18nProvider>
-    </ThemeProvider>
+    <UserVarProvider userVar={currentUserVar}>
+      <ThemeProvider theme={canada}>
+        <I18nProvider i18n={i18n}>
+          <Router>
+            <App />
+          </Router>
+        </I18nProvider>
+      </ThemeProvider>
+    </UserVarProvider>
   </ApolloProvider>,
   document.getElementById('root'),
 )
